refactor(spp): split circuit.go and proof_bundle.go into focused files
Pure file/function organization, no behavior change. circuit.go now
reads as the protocol flow (structs, NewCircuit, Define); generic
constraint helpers move to constraints.go and the spend checks to
spend.go. proof_bundle.go keeps only the bundle entry points, with
witness assembly in proof_assignment.go, request parsing in
proof_parse.go, public-amount handling in public_amount.go, and the
external-data hasher (with its struct) in external_data.go. The
off-circuit PublicInputs type and PublicInputHash join the public-input
names list in public_inputs.go.
