+++ prover/server/prover/spp/circuit.go
@@ -2,7 +2,6 @@ package spp
 
 import (
 	"fmt"
-	"math/big"
 
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/std/algebra/emulated/sw_emulated"
@@ -113,112 +112,6 @@ func MustNewCircuit(shape Shape) *Circuit {
 	return circuit
 }
 
-// assertEqualWhen constrains a == b only when cond == 1 (cond must be boolean).
-func assertEqualWhen(api frontend.API, cond, a, b frontend.Variable) {
-	api.AssertIsEqual(api.Mul(cond, api.Sub(a, b)), 0)
-}
-
-// assertZeroWhen constrains v == 0 only when cond == 1 (cond must be boolean).
-func assertZeroWhen(api frontend.API, cond, v frontend.Variable) {
-	api.AssertIsEqual(api.Mul(cond, v), 0)
-}
-
-// assertStrictlyOrdered constrains lo < mid < hi for a real entry; dummy entries
-// (isDummy == 1) are mapped to 0 < 1 < 2 so the check always holds for them.
-// Expressed with AssertIsLessOrEqual + AssertIsDifferent rather than a `+1`
-// increment, which could wrap at the field boundary.
-func assertStrictlyOrdered(api frontend.API, isDummy, lo, mid, hi frontend.Variable) {
-	lo = api.Select(isDummy, frontend.Variable(0), lo)
-	mid = api.Select(isDummy, frontend.Variable(1), mid)
-	hi = api.Select(isDummy, frontend.Variable(2), hi)
-	api.AssertIsLessOrEqual(lo, mid)
-	api.AssertIsDifferent(lo, mid)
-	api.AssertIsLessOrEqual(mid, hi)
-	api.AssertIsDifferent(mid, hi)
-}
-
-// spendEnv holds the values shared across every input-spend check: they are
-// derived once per proof from the wallet secret and the single P256 signer.
-type spendEnv struct {
-	nullifierPkFromSecret frontend.Variable
-	p256OwnerKeyHash      frontend.Variable
-	p256SigValid          frontend.Variable
-	nullifierSecret       frontend.Variable
-}
-
-// constrainInput verifies one spent input — domain, state-tree inclusion, owner
-// binding, nullifier derivation, and nullifier-tree non-inclusion — and returns
-// its UTXO hash (0 for a dummy) for the transaction-hash chain.
-func constrainInput(api frontend.API, in Input, env spendEnv) frontend.Variable {
-	api.AssertIsBoolean(in.IsDummy)
-	notDummy := api.Sub(1, in.IsDummy)
-
-	assertZeroWhen(api, in.IsDummy, in.Utxo.AssetAmount)
-	assertEqualWhen(api, notDummy, in.Utxo.Domain, UtxoDomain) // pin domain
-
-	utxoHash := UtxoHashCircuit(api, in.Utxo)
-
-	// Inclusion: utxoHash is a leaf of the state tree at UtxoTreeRoot.
-	stateRoot := StatePathFoldCircuit(api, utxoHash, in.State.Siblings, in.State.Directions)
-	assertEqualWhen(api, notDummy, stateRoot, in.UtxoTreeRoot)
-
-	// Owner binding: P256 inputs (SolanaPkHash == 0) recompute the owner key hash
-	// from the witnessed P256 point; Solana inputs use the public hash.
-	isP256 := api.IsZero(in.SolanaPkHash)
-	ownerKeyHash := api.Select(isP256, env.p256OwnerKeyHash, in.SolanaPkHash)
-	ownerHash := OwnerHashCircuit(api, ownerKeyHash, in.NullifierPk)
-	assertEqualWhen(api, notDummy, ownerHash, in.Utxo.Owner)
-	assertEqualWhen(api, notDummy, env.nullifierPkFromSecret, in.NullifierPk)
-	// Real P256 inputs must carry a valid signature; Solana inputs are verified
-	// by SPP out of circuit.
-	assertZeroWhen(api, api.Mul(notDummy, isP256), api.Sub(1, env.p256SigValid))
-	assertZeroWhen(api, in.IsDummy, in.NullifierPk)
-	assertZeroWhen(api, in.IsDummy, in.SolanaPkHash)
-
-	// Nullifier: derived from the UTXO hash, blinding, and shared secret.
-	nullifier := NullifierHashCircuit(api, utxoHash, in.Utxo.Blinding, env.nullifierSecret)
-	assertEqualWhen(api, notDummy, nullifier, in.Nullifier)
-	assertZeroWhen(api, in.IsDummy, in.Nullifier)
-
-	// Non-inclusion: the low leaf is in the nullifier tree and brackets the
-	// nullifier (NfLowValue < Nullifier < NfNextValue).
-	lowLeaf := IndexedLeafHashCircuit(api, in.NfLowValue, in.NfNextValue)
-	nfRoot := StatePathFoldCircuit(api, lowLeaf, in.NfLow.Siblings, in.NfLow.Directions)
-	assertEqualWhen(api, notDummy, nfRoot, in.NullifierRoot)
-	assertStrictlyOrdered(api, in.IsDummy, in.NfLowValue, in.Nullifier, in.NfNextValue)
-
-	return api.Select(in.IsDummy, frontend.Variable(0), utxoHash)
-}
-
-// constrainOutput verifies one created output and returns its UTXO hash (0 for a
-// dummy) for the transaction-hash chain.
-func constrainOutput(api frontend.API, out Output) frontend.Variable {
-	api.AssertIsBoolean(out.IsDummy)
-	notDummy := api.Sub(1, out.IsDummy)
-
-	assertZeroWhen(api, out.IsDummy, out.Utxo.AssetAmount)
-	assertEqualWhen(api, notDummy, out.Utxo.Domain, UtxoDomain) // pin domain
-
-	utxoHash := UtxoHashCircuit(api, out.Utxo)
-	assertEqualWhen(api, notDummy, utxoHash, out.Hash)
-	assertZeroWhen(api, out.IsDummy, out.Hash)
-
-	return api.Select(out.IsDummy, frontend.Variable(0), utxoHash)
-}
-
-// assertDistinctNullifiers rejects spending the same input twice in one
-// transaction: every pair of real inputs must carry distinct nullifiers. Dummy
-// inputs all carry nullifier 0 and are excluded.
-func (c *Circuit) assertDistinctNullifiers(api frontend.API) {
-	for i := range c.Inputs {
-		for j := i + 1; j < len(c.Inputs); j++ {
-			bothReal := api.Mul(api.Sub(1, c.Inputs[i].IsDummy), api.Sub(1, c.Inputs[j].IsDummy))
-			sameNullifier := api.IsZero(api.Sub(c.Inputs[i].Nullifier, c.Inputs[j].Nullifier))
-			api.AssertIsEqual(api.Mul(bothReal, sameNullifier), 0)
-		}
-	}
-}
-
 func (c *Circuit) Define(api frontend.API) error {
 	if err := c.validateShape(); err != nil {
 		return err
@@ -324,56 +217,4 @@ func (c *Circuit) validateShape() error {
 		}
 	}
 	return nil
-}
-
-type PublicInputs struct {
-	Nullifiers           []*big.Int
-	OutputUtxoHashes     []*big.Int
-	UtxoTreeRoots        []*big.Int
-	NullifierRoots       []*big.Int
-	PrivateTxHash        *big.Int
-	ExternalDataHash     *big.Int
-	PublicSolAmount      *big.Int
-	PublicSplAmount      *big.Int
-	PublicSplAssetPubkey *big.Int
-	ProgramIDHashChain   *big.Int
-	SolanaPubkeyHash     *big.Int
-	SolanaPkHashes       []*big.Int
-}
-
-func PublicInputHash(inputs PublicInputs) (*big.Int, error) {
-	nullifierChain, err := HashChain(inputs.Nullifiers)
-	if err != nil {
-		return nil, fmt.Errorf("spp: public input hash nullifier chain: %w", err)
-	}
-	outputChain, err := HashChain(inputs.OutputUtxoHashes)
-	if err != nil {
-		return nil, fmt.Errorf("spp: public input hash output chain: %w", err)
-	}
-	utxoRootChain, err := HashChain(inputs.UtxoTreeRoots)
-	if err != nil {
-		return nil, fmt.Errorf("spp: public input hash UTXO root chain: %w", err)
-	}
-	nullifierRootChain, err := HashChain(inputs.NullifierRoots)
-	if err != nil {
-		return nil, fmt.Errorf("spp: public input hash nullifier root chain: %w", err)
-	}
-	solanaOwnerKeyHashChain, err := HashChain(inputs.SolanaPkHashes)
-	if err != nil {
-		return nil, fmt.Errorf("spp: public input hash solana pk hash chain: %w", err)
-	}
-	return HashChain([]*big.Int{
-		nullifierChain,
-		outputChain,
-		utxoRootChain,
-		nullifierRootChain,
-		inputs.PrivateTxHash,
-		inputs.ExternalDataHash,
-		inputs.PublicSolAmount,
-		inputs.PublicSplAmount,
-		inputs.PublicSplAssetPubkey,
-		inputs.ProgramIDHashChain,
-		inputs.SolanaPubkeyHash,
-		solanaOwnerKeyHashChain,
-	})
 }
